response updated

+++ client/_IRCClient.hpp
@@ -12,7 +12,7 @@ class _IRCClient {
 
 		void parse_input();
 		void reg(); //register is a reserved keyword
-		response_pair response(uint16_t r, ...);
+		response_pair response(int r, ...);
 
 		std::string password;
 		std::string nickname;
